Split test_pubmed from test_citations
